% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam.R
\name{jam}
\alias{jam}
\alias{unjam}
\title{Serialize R objects to files}
\usage{
jam(obj, file = sprintf("./data/\%s.jam", deparse(substitute(obj))))

unjam(file)
}
\arguments{
\item{obj}{atomic vector or list, with or without attributes}

\item{file}{archive file name. Defaults to "./data/[obj_name].jam"}
}
\description{
Fast serialization of R objects into binary archives. All types of atomic
vectors, arbitrarly nested lists and data frames are supported. Atributes of
supported data types are also saved except of the row.names. Non suported
attribute types are silently dropped.
}
\examples{
\dontrun{
  jam(iris, "./data/iris.jam")
  all.equal(iris, unjam("./data/iris.jam"))
}
}
\author{
Vitalie Spinu
}

